k8s: Move CastInformerEvent to informer
Move CastInformerEvent function to k8s/informer. Users of this function
already import k8s/informer, and this allows some of them to remove
import of 'k8s' itself.
