+++ pkg/k8s/factory_functions.go
@@ -15,31 +15,7 @@ import (
 	slim_corev1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/api/core/v1"
 	slim_metav1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/apis/meta/v1"
 	"github.com/cilium/cilium/pkg/k8s/types"
-	"github.com/cilium/cilium/pkg/logging/logfields"
 )
-
-// CastInformerEvent tries to cast obj to type typ, directly
-// or by DeletedFinalStateUnknown type. It returns nil and logs
-// an error if obj doesn't contain type typ.
-func CastInformerEvent[typ any](obj interface{}) *typ {
-	k8sObj, ok := obj.(*typ)
-	if ok {
-		return k8sObj
-	}
-	deletedObj, ok := obj.(cache.DeletedFinalStateUnknown)
-	if ok {
-		// Delete was not observed by the watcher but is
-		// removed from kube-apiserver. This is the last
-		// known state and the object no longer exists.
-		k8sObj, ok := deletedObj.Obj.(*typ)
-		if ok {
-			return k8sObj
-		}
-	}
-	log.WithField(logfields.Object, logfields.Repr(obj)).
-		Warnf("Ignoring invalid type, expected: %T", new(typ))
-	return nil
-}
 
 // AnnotationsEqual returns whether the annotation with any key in
 // relevantAnnotations is equal in anno1 and anno2.
