+++ pkg/k8s/informer/benchmarks/informer_benchmarks_test.go
@@ -19,7 +19,6 @@ import (
 	"k8s.io/client-go/tools/cache"
 
 	"github.com/cilium/cilium/pkg/annotation"
-	"github.com/cilium/cilium/pkg/k8s"
 	slim_corev1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/api/core/v1"
 
 	"github.com/cilium/cilium/pkg/k8s/informer"
@@ -430,8 +429,8 @@ func benchmarkInformer(ctx context.Context, nCycles int, newInformer bool, b *te
 			cache.ResourceEventHandlerFuncs{
 				AddFunc: func(obj interface{}) {},
 				UpdateFunc: func(oldObj, newObj interface{}) {
-					if oldK8sNP := k8s.CastInformerEvent[slim_corev1.Node](oldObj); oldK8sNP != nil {
-						if newK8sNP := k8s.CastInformerEvent[slim_corev1.Node](newObj); newK8sNP != nil {
+					if oldK8sNP := informer.CastInformerEvent[slim_corev1.Node](oldObj); oldK8sNP != nil {
+						if newK8sNP := informer.CastInformerEvent[slim_corev1.Node](newObj); newK8sNP != nil {
 							if reflect.DeepEqual(oldK8sNP, newK8sNP) {
 								return
 							}
@@ -439,7 +438,7 @@ func benchmarkInformer(ctx context.Context, nCycles int, newInformer bool, b *te
 					}
 				},
 				DeleteFunc: func(obj interface{}) {
-					k8sNP := k8s.CastInformerEvent[slim_corev1.Node](obj)
+					k8sNP := informer.CastInformerEvent[slim_corev1.Node](obj)
 					if k8sNP == nil {
 						deletedObj, ok := obj.(cache.DeletedFinalStateUnknown)
 						if !ok {
@@ -448,7 +447,7 @@ func benchmarkInformer(ctx context.Context, nCycles int, newInformer bool, b *te
 						// Delete was not observed by the watcher but is
 						// removed from kube-apiserver. This is the last
 						// known state and the object no longer exists.
-						k8sNP = k8s.CastInformerEvent[slim_corev1.Node](deletedObj.Obj)
+						k8sNP = informer.CastInformerEvent[slim_corev1.Node](deletedObj.Obj)
 						if k8sNP == nil {
 							return
 						}
