+++ pkg/k8s/watchers/pod.go
@@ -166,7 +166,7 @@ func (k *K8sPodWatcher) createAllPodsController(slimClient slimclientset.Interfa
 		0,
 		cache.ResourceEventHandlerFuncs{
 			AddFunc: func(obj interface{}) {
-				if pod := k8s.CastInformerEvent[slim_corev1.Pod](obj); pod != nil {
+				if pod := informer.CastInformerEvent[slim_corev1.Pod](obj); pod != nil {
 					err := k.addK8sPodV1(pod)
 					k.k8sEventReporter.K8sEventProcessed(metricPod, resources.MetricCreate, err == nil)
 					k.k8sEventReporter.K8sEventReceived(podApiGroup, metricPod, resources.MetricCreate, true, false)
@@ -175,8 +175,8 @@ func (k *K8sPodWatcher) createAllPodsController(slimClient slimclientset.Interfa
 				}
 			},
 			UpdateFunc: func(oldObj, newObj interface{}) {
-				if oldPod := k8s.CastInformerEvent[slim_corev1.Pod](oldObj); oldPod != nil {
-					if newPod := k8s.CastInformerEvent[slim_corev1.Pod](newObj); newPod != nil {
+				if oldPod := informer.CastInformerEvent[slim_corev1.Pod](oldObj); oldPod != nil {
+					if newPod := informer.CastInformerEvent[slim_corev1.Pod](newObj); newPod != nil {
 						if oldPod.DeepEqual(newPod) {
 							k.k8sEventReporter.K8sEventReceived(podApiGroup, metricPod, resources.MetricUpdate, false, true)
 						} else {
@@ -190,7 +190,7 @@ func (k *K8sPodWatcher) createAllPodsController(slimClient slimclientset.Interfa
 				}
 			},
 			DeleteFunc: func(obj interface{}) {
-				if pod := k8s.CastInformerEvent[slim_corev1.Pod](obj); pod != nil {
+				if pod := informer.CastInformerEvent[slim_corev1.Pod](obj); pod != nil {
 					err := k.deleteK8sPodV1(pod)
 					k.k8sEventReporter.K8sEventProcessed(metricPod, resources.MetricDelete, err == nil)
 					k.k8sEventReporter.K8sEventReceived(podApiGroup, metricPod, resources.MetricDelete, true, false)
